+++ handler/rtu.py
@@ -781,7 +781,7 @@ class RtuCtlHandler(base.RequestHandler):
         self.finish()
         if env:
             cur = yield self.write_event(
-                event_id, contents, 2, user_name=user_data['user_name'])
+                event_id, contents, 2, user_name=user_data['user_name'],app_unique=rqmsg.head.unique)
         del msg, rqmsg, user_data, user_uuid
 
 
@@ -915,5 +915,5 @@ class RtuTimerCtlHandler(base.RequestHandler):
         self.write(mx.code_pb2(msg, self._go_back_format))
         self.finish()
         if env and rqmsg.data_mark == 1:
-            self.write_event(11, contents, 2, user_name=user_data['user_name'])
+            self.write_event(11, contents, 2, user_name=user_data['user_name'],app_unique=rqmsg.head.unique)
         del msg, rqmsg, user_data